重新组织代码分布
