+++ kernel/src/net/net_core.rs
@@ -145,99 +145,3 @@ pub fn poll_ifaces() {
         iface.poll();
     }
 }
-
-// /// 对ifaces进行轮询，最多对SOCKET_SET尝试times次加锁。
-// ///
-// /// @return 轮询成功，返回Ok(())
-// /// @return 加锁超时，返回SystemError::EAGAIN_OR_EWOULDBLOCK
-// /// @return 没有网卡，返回SystemError::ENODEV
-// pub fn poll_ifaces_try_lock(times: u16) -> Result<(), SystemError> {
-//     let mut i = 0;
-//     while i < times {
-//         let guard: RwLockReadGuard<BTreeMap<usize, Arc<dyn Iface>>> =
-//             NET_DEVICES.read_irqsave();
-//         if guard.len() == 0 {
-//             warn!("poll_ifaces: No net driver found!");
-//             // 没有网卡，返回错误
-//             return Err(SystemError::ENODEV);
-//         }
-//         for (_, iface) in guard.iter() {
-//             iface.poll();
-//         }
-//         return Ok(());
-//     }
-//     // 尝试次数用完，返回错误
-//     return Err(SystemError::EAGAIN_OR_EWOULDBLOCK);
-// }
-
-// /// 对ifaces进行轮询，最多对SOCKET_SET尝试一次加锁。
-// ///
-// /// @return 轮询成功，返回Ok(())
-// /// @return 加锁超时，返回SystemError::EAGAIN_OR_EWOULDBLOCK
-// /// @return 没有网卡，返回SystemError::ENODEV
-// pub fn poll_ifaces_try_lock_onetime() -> Result<(), SystemError> {
-//     let guard: RwLockReadGuard<BTreeMap<usize, Arc<dyn Iface>>> = NET_DEVICES.read_irqsave();
-//     if guard.len() == 0 {
-//         warn!("poll_ifaces: No net driver found!");
-//         // 没有网卡，返回错误
-//         return Err(SystemError::ENODEV);
-//     }
-//     for (_, iface) in guard.iter() {
-//         let _ = iface.poll();
-//     }
-//     send_event()?;
-//     return Ok(());
-// }
-
-// /// ### 处理轮询后的事件
-// fn send_event() -> Result<(), SystemError> {
-//     for (handle, socket_type) in .lock().iter() {
-
-//         let global_handle = GlobalSocketHandle::new_smoltcp_handle(handle);
-
-//         let handle_guard = HANDLE_MAP.read_irqsave();
-//         let item: Option<&super::socket::SocketHandleItem> = handle_guard.get(&global_handle);
-//         if item.is_none() {
-//             continue;
-//         }
-
-//         let handle_item = item.unwrap();
-//         let posix_item = handle_item.posix_item();
-//         if posix_item.is_none() {
-//             continue;
-//         }
-//         let posix_item = posix_item.unwrap();
-
-//         // 获取socket上的事件
-//         let mut events = SocketPollMethod::poll(socket_type, handle_item).bits() as u64;
-
-//         // 分发到相应类型socket处理
-//         match socket_type {
-//             smoltcp::socket::Socket::Raw(_) | smoltcp::socket::Socket::Udp(_) => {
-//                 posix_item.wakeup_any(events);
-//             }
-//             smoltcp::socket::Socket::Icmp(_) => unimplemented!("Icmp socket hasn't unimplemented"),
-//             smoltcp::socket::Socket::Tcp(inner_socket) => {
-//                 if inner_socket.is_active() {
-//                     events |= TcpSocket::CAN_ACCPET;
-//                 }
-//                 if inner_socket.state() == smoltcp::socket::tcp::State::Established {
-//                     events |= TcpSocket::CAN_CONNECT;
-//                 }
-//                 if inner_socket.state() == smoltcp::socket::tcp::State::CloseWait {
-//                     events |= EPollEventType::EPOLLHUP.bits() as u64;
-//                 }
-
-//                 posix_item.wakeup_any(events);
-//             }
-//             smoltcp::socket::Socket::Dhcpv4(_) => {}
-//             smoltcp::socket::Socket::Dns(_) => unimplemented!("Dns socket hasn't unimplemented"),
-//         }
-//         EventPoll::wakeup_epoll(
-//             &posix_item.epitems,
-//             EPollEventType::from_bits_truncate(events as u32),
-//         )?;
-//         drop(handle_guard);
-//     }
-//     Ok(())
-// }
