+++ kernel/src/net/socket/common/mod.rs
@@ -1,10 +1,8 @@
 // pub mod poll_unit;
 mod epoll_items;
-pub mod shutdown;
 
+pub mod shutdown;
 pub use epoll_items::EPollItems;
-#[allow(dead_code)]
-pub use shutdown::Shutdown;
 
 // /// @brief 在trait Socket的metadata函数中返回该结构体供外部使用
 // #[derive(Debug, Clone)]
