+++ kernel/src/net/socket/mod.rs
@@ -1,7 +1,7 @@
 mod base;
 mod buffer;
 mod common;
-mod define;
+mod definition;
 mod endpoint;
 mod family;
 pub mod inet;
@@ -18,7 +18,7 @@ pub use common::{
     // poll_unit::{EPollItems, WaitQueue},
     EPollItems,
 };
-pub use define::*;
+pub use definition::*;
 pub use endpoint::*;
 pub use family::{AddressFamily, Family};
 pub use inode::Inode;
